moved client utils to utils.go
