+++ client/client.go
@@ -10,8 +10,6 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/azidentity"
 	"github.com/microsoft/kiota-abstractions-go/serialization"
 	msgraphsdk "github.com/microsoftgraph/msgraph-sdk-go"
-	"github.com/microsoftgraph/msgraph-sdk-go/models"
-	"github.com/microsoftgraph/msgraph-sdk-go/models/odataerrors"
 	"github.com/microsoftgraph/msgraph-sdk-go/solutions"
 )
 
@@ -195,94 +193,5 @@ func newBookingCalendarViewRequest(start string, end string) *solutions.BookingB
 
 	return &solutions.BookingBusinessesItemCalendarViewRequestBuilderGetRequestConfiguration{
 		QueryParameters: qp,
-	}
-}
-
-func getDateTime(dt models.DateTimeTimeZoneable) (time.Time, error) {
-	dateTime := dt.GetDateTime()
-	if dateTime == nil {
-		return time.Time{}, fmt.Errorf("DateTime is required")
-	}
-	return time.Parse(time.RFC3339Nano, *dateTime)
-}
-
-func getCustomerValue(a models.BookingAppointmentable) BookingCustomer {
-	return BookingCustomer{
-		Name:         getStringValue(a.GetCustomerName()),
-		EmailAddress: getStringValue(a.GetCustomerEmailAddress()),
-		Phone:        getStringValue(a.GetCustomerPhone()),
-		Notes:        getStringValue(a.GetCustomerNotes()),
-		TimeZone:     getStringValue(a.GetCustomerTimeZone()),
-	}
-}
-
-func getQuestionsFromStore(c models.BookingCustomerInformationBaseable) []BookingQuestion {
-	questions, err := c.GetBackingStore().Get("customQuestionAnswers")
-	if questions == nil || err != nil {
-		return nil
-	}
-
-	q := make([]BookingQuestion, 0)
-
-	for _, question := range questions.([]models.BookingQuestionAnswerable) {
-		q = append(q, BookingQuestion{
-			QuestionID:   *question.GetQuestionId(),
-			IsRequired:   *question.GetIsRequired(),
-			QuestionText: *question.GetQuestion(),
-			Answer:       *question.GetAnswer(),
-		})
-	}
-
-	return q
-}
-
-func getStringFromStore(c models.BookingCustomerInformationBaseable, key string) *string {
-	val, ok := c.GetBackingStore().Get(key)
-	if ok != nil {
-		log.Fatal("Error getting value from store: ", ok)
-	}
-
-	return val.(*string)
-}
-
-func getStringValue(s *string) string {
-	if s == nil {
-		return ""
-	}
-	return *s
-}
-
-func getTimeValue(t *time.Time) time.Time {
-	if t == nil {
-		return time.Time{}
-	}
-	return *t
-}
-
-func getIntValue(i *int32) int32 {
-	if i == nil {
-		return 0
-	}
-	return *i
-}
-
-func getDurationValue(d *serialization.ISODuration) serialization.ISODuration {
-	if d == nil {
-		return serialization.ISODuration{}
-	}
-	return *d
-}
-
-func printOdataError(err error) {
-	switch err.(type) {
-	case *odataerrors.ODataError:
-		typed := err.(*odataerrors.ODataError)
-		fmt.Printf("error:", typed.Error())
-		if terr := typed.GetErrorEscaped(); terr != nil {
-			fmt.Printf("code: %s", *terr.GetCode())
-			fmt.Printf("msg: %s", *terr.GetMessage())
-		}
-	default:
-		fmt.Printf("%T > error: %#v", err, err)
 	}
 }
